MODIFIED : MainRight.scss 수정
