+++ public/data/MockData.json
@@ -1,0 +1,20 @@
+[
+  {
+    "article_number": 1,
+    "myprofile": "/public/images/41_dagyeong/myProfile.jpeg",
+    "posting": "/public/images/41_dagyeong/pic2.jpeg",
+    "like": "좋아요 3270개",
+    "comment_nickname": "nana123",
+    "comment": "위워크에서 진행한 필름카메라 클래스...",
+    "time": "35분전"
+  },
+  {
+    "article_number": 2,
+    "myprofile": "/public/images/41_dagyeong/myProfile.jpeg",
+    "posting": "/public/images/41_dagyeong/pic3.jpeg",
+    "like": "좋아요 45개",
+    "comment_nickname": "squre365",
+    "comment": "오늘 저녁은 뭐 먹지?",
+    "time": "16분전"
+  }
+]
