+++ src/pages/41_dagyeong/Main/Components/Article.scss
@@ -1,0 +1,108 @@
+@import '/src/pages/41_dagyeong/styles/mixin.scss';
+@import '/src/pages/41_dagyeong/styles/variables.scss';
+
+article {
+  width: 470px;
+  // height: 750px;
+  border: 1px solid #e6e6e6;
+  border-radius: 5px;
+
+  .feedTop {
+    height: 100%;
+    display: flex;
+    align-items: center;
+    height: 50px;
+
+    .feedprofile_img {
+      width: 50px;
+      height: 50px;
+      border-radius: 70%;
+    }
+
+    .feeds_nickName {
+      font-weight: bold;
+      margin-left: 5px;
+    }
+
+    .feedTop_option {
+      margin-left: 330px;
+      width: 32px;
+      height: 32px;
+    }
+  }
+
+  .feeds_img {
+    width: 468px;
+    height: 470px;
+  }
+
+  .article_icons > img {
+    width: 28px;
+    height: 28px;
+  }
+
+  .article_icons img:nth-child(1) {
+    margin-left: 5px;
+  }
+
+  .article_icons img:nth-child(-n + 3) {
+    margin-right: 5px;
+  }
+
+  .article_icons img:nth-child(4) {
+    margin-left: 332px;
+  }
+
+  .like {
+    margin-left: 10px;
+    font-size: 14px;
+  }
+
+  .commentLists {
+    .comment_ul {
+      list-style: none;
+      padding: 0;
+      // margin-left: 10px;
+      margin: 10px 10px;
+      display: flex;
+      flex-direction: column;
+
+      .nickName {
+        font-weight: bold;
+      }
+
+      .comment {
+        margin-left: 10px;
+      }
+
+      .delete {
+        margin-left: 10px;
+      }
+    }
+  }
+  .comment_box {
+    display: flex;
+    margin-top: 0;
+    // flex-direction: column;
+  }
+
+  hr {
+    background: #e6e6e6;
+    height: 1px;
+    border: 0;
+  }
+
+  .do_comment > input {
+    width: 420px;
+    height: 30px;
+    border: none;
+    text-indent: 10px;
+    margin-bottom: 5px;
+  }
+
+  .do_comment > button {
+    border: none;
+    background-color: transparent;
+    color: #c5e2fb;
+  }
+}
