+++ src/pages/41_dagyeong/Main/Components/MainRight.scss
@@ -1,0 +1,110 @@
+@import '/src/pages/41_dagyeong/styles/mixin.scss';
+@import '/src/pages/41_dagyeong/styles/variables.scss';
+
+/* main_right */
+.main_right {
+  margin-left: 15px;
+
+  .myprofile {
+    display: flex;
+
+    .myprofile span:nth-child(1) {
+      font-weight: bold;
+    }
+    .my_profile {
+      display: flex;
+      flex-direction: column;
+      margin-left: 10px;
+    }
+    .myprofile_img {
+      width: 50px;
+      height: 50px;
+      border-radius: 70%;
+    }
+  }
+}
+
+.feedprofile_img {
+  width: 40px;
+  height: 40px;
+  border-radius: 70%;
+  padding: 3px;
+}
+
+/* 스토리 */
+
+.storybox {
+  border: 1px solid #e6e6e6;
+  width: 380px;
+  height: 185px;
+  margin-bottom: 20px;
+  margin-top: 10px;
+  border-radius: 3px;
+
+  .storybox_top {
+    padding: 10px;
+    color: #b9b9b9;
+  }
+
+  .story_all_view {
+    margin-left: 252px;
+    font-weight: bold;
+    color: black;
+  }
+}
+.story_nickname {
+  display: flex;
+  margin-left: 10px;
+  margin-bottom: 8px;
+
+  .story_state {
+    display: flex;
+    flex-direction: column;
+    margin-left: 10px;
+
+    .story_time {
+      color: #b9b9b9;
+    }
+  }
+}
+
+.story_anotherpeople {
+  margin: 0;
+  color: #b9b9b9;
+}
+
+.story_state p {
+  margin: 0;
+}
+
+.recommend {
+  //1
+  border: 1px solid #e6e6e6;
+  width: 380px;
+  height: 185px;
+  border-radius: 3px;
+
+  .recommend_top {
+    padding: 10px;
+    color: #b9b9b9;
+
+    .recommend_all_view {
+      margin-left: 175px;
+      font-weight: bold;
+      color: black;
+    }
+  }
+
+  .recommend_follow > button {
+    border: none;
+    background-color: transparent;
+    margin-top: 15px;
+    color: #56a5f2;
+  }
+}
+
+.aside {
+  width: 380px;
+  color: #b9b9b9;
+  margin-top: 20px;
+}
